vmm/kvm: Selectively send the 'cont' command depending on the KVM version
Sending the 'cont' command after loading a VM causes a race
condition and hangs the VM with qemu-kvm 0.11.0 and newer.
With these versions of KVM, the virtual CPUs of a VM are started
automatically after loading its image, and thus sending 'cont'
is not necessary. The KVM driver is changed to judge whether
to send the command based on the KVM version, so that it is not
issued when unnecessary to avoid the race condition.