vmm/kvm: Add a workaround for the slow 'migrate' issue with KVM
With qemu-kvm 0.11.0 and newer, the 'migrate' command is very slow
when simply "exec:dd of=..." is passed as its argument, as has been
done in the KVM driver. The command passed to 'migrate' is changed
to include a pipe (i.e., "exec:dd bs=1M | dd bs=1M of=...) to
circumvent this problem. This workaround follows what libvirt does,
although somewhat implicitly, and thus is adapted in the driver.
The exact reason this piping helps has not been tracked down yet.

Also, "bs=1M" is added as an argument to the 'dd' command, so that
it uses large blocks for better performance.