vmm/virtualbox: Allow 64-bit guests if requested in virtualbox.setup
Add "64-bit" property to virtualbox.setup which, if true, requests that
the machine be configured for a 64-bit guest.  This causes the following
configuration changes:

* VT is enabled for the machine (VBox requirement)
* IO-APIC is enabled (commonly required by 64-bit guests)
* LM CPUID flag is enabled
* OS type is set to Linux26_64 to inform VBox that a 64-bit guest is
present.  This will enable 64-bit support on 32-bit hosts.

VBox will display a warning dialog when attempting to resume a 64-bit
parcel on a host without the requisite support.