tools/dirtometer: Simplify window-sizing calculations
- Remove code that attempted to calculate an optimal size for the chunk
bitmap, which is no longer necessary now that we have a scroll bar.
Instead, configure a reasonable size request for the chunk bitmap and
allow gtk's allocation process to handle the heavy lifting.

- If the chunk bitmap pane is not enabled, disable window resizing and
force the window to the minimum size

- Use img_viewport->allocation.width to calculate the viewport width in
update_img().  Because we are overriding the viewport's size request, it
appears that we no longer receive a size-allocate signal when starting
with the default size.

- Eliminate pane size-request and size-allocate signal handlers, since
they are no longer necessary