tools/viewer: Work around unfullscreen problem with Metacity
- If the viewer starts in full-screen mode and then the user unfullscreens
with RightCtrl-F, there was a period of time in which the window was
unfullscreened but was still sized and placed to cover the entire screen.
Metacity would notice this and "helpfully" return the window to
full-screen mode, making it impossible to unfullscreen the viewer.  Work
around this by making an extra resize call before unfullscreening.